% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinAdditionalSpecies.R
\name{joinAdditionalSpecies}
\alias{joinAdditionalSpecies}
\title{joinAdditionalSpecies: compiles additional species data}
\usage{
joinAdditionalSpecies(
  park = "all",
  from = 2007,
  to = 2021,
  QAQC = FALSE,
  panels = 1:4,
  locType = c("VS", "all"),
  eventType = c("complete", "all"),
  speciesType = c("all", "native", "exotic", "invasive")
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"APCO"}{Appomattox Court House NHP only}
\item{"ASIS"}{Assateague Island National Seashore}
\item{"BOWA"}{Booker T. Washington NM only}
\item{"COLO"}{Colonial NHP only}
\item{"FRSP"}{Fredericksburg & Spotsylvania NMP only}
\item{"GETT"}{Gettysburg NMP only}
\item{"GEWA"}{George Washington Birthplace NM only}
\item{"HOFU"}{Hopewell Furnace NHS only}
\item{"PETE"}{Petersburg NBP only}
\item{"RICH"}{Richmond NB only}
\item{"SAHI"}{Sagamore Hill NHS only}
\item{"THST"}{Thomas Stone NHS only}
\item{"VAFO"}{Valley Forge NHP only}}}

\item{from}{Year to start analysis, ranging from 2007 to current year}

\item{to}{Year to stop analysis, ranging from 2007 to current year}

\item{QAQC}{Allows you to remove or include QAQC events.
\describe{
\item{FALSE}{Default. Only returns visits that are not QAQC visits}
\item{TRUE}{Returns all visits, including QAQC visits}}}

\item{panels}{Allows you to select individual panels from 1 to 4. Default is all 4 panels (1:4).
If more than one panel is selected, specify by c(1, 3), for example.}

\item{locType}{Allows you to only include plots that are part of the GRTS sample design or
include all plots, such as deer exclosures.
\describe{
\item{"VS"}{Only include plots that are part of the Vital Signs GRTS sample design}
\item{"all"}{Include all plots, such as plots in deer exclosures or test plots.}}}

\item{eventType}{Allows you to include only complete sampling events or all sampling events
\describe{
\item{"complete"}{Default. Only include sampling events for a plot that are complete.}
\item{"all}{Include all plot events with a record in tblCOMN.Event, including plots missing most of the data
associated with that event (eg COLO-380-2018). This feature is currently hard-coded in the function.}}}

\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
\item{"invasive"}{Returns species on the Indicator Invasive List}
}}
}
\value{
Returns a dataframe with a row for each species recorded during the timed search per visit.
}
\description{
This function compiles the additional species collected during a timed 15-minute plot search. This is
primarily an internal function used for QAQC and to generate the species lists for makeSppList(). Note that the timed
search wasn't implemented until 2009, and only included species on the MIDN indicator list, which has also had additions
over time. Sampleing in 2019, all woody species were included in the timed search.For more information about how the
indicator list and protocol had changed ovre time, refer to the Summary of Major Protocol Changes document
for the MIDN forest protocol located in the Long-Term Forest Monitoring Protocol IRMA Project:
   https://irma.nps.gov/Datastore/Reference/Profile/2189101.
}
\examples{
\dontrun{
importData()
# compile invasive species found in plot search in VAFO for all years
SARA_quads <- joinAdditionalSpecies(park = 'VAFO', speciesType = 'invasive')

# compile all species, including QAQC visits for parks in cycle 3
native_quads <- joinQuadSpecies(from = 2015, to = 2018, QAQC = TRUE)
}

}
