% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinMicroSaplings.R
\name{joinMicroSaplings}
\alias{joinMicroSaplings}
\title{joinMicroSaplings: compiles sapling data collected in microplots}
\usage{
joinMicroSaplings(
  park = "all",
  from = 2007,
  to = 2021,
  QAQC = FALSE,
  panels = 1:4,
  locType = c("VS", "all"),
  eventType = c("complete", "all"),
  status = c("active", "live", "dead", "all"),
  speciesType = c("all", "native", "exotic", "invasive"),
  canopyForm = c("all", "canopy"),
  ...
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"APCO"}{Appomattox Court House NHP only}
\item{"ASIS"}{Assateague Island National Seashore}
\item{"BOWA"}{Booker T. Washington NM only}
\item{"COLO"}{Colonial NHP only}
\item{"FRSP"}{Fredericksburg & Spotsylvania NMP only}
\item{"GETT"}{Gettysburg NMP only}
\item{"GEWA"}{George Washington Birthplace NM only}
\item{"HOFU"}{Hopewell Furnace NHS only}
\item{"PETE"}{Petersburg NBP only}
\item{"RICH"}{Richmond NB only}
\item{"SAHI"}{Sagamore Hill NHS only}
\item{"THST"}{Thomas Stone NHS only}
\item{"VAFO"}{Valley Forge NHP only}}}

\item{from}{Year to start analysis, ranging from 2007 to current year}

\item{to}{Year to stop analysis, ranging from 2007 to current year}

\item{QAQC}{Allows you to remove or include QAQC events.
\describe{
\item{FALSE}{Default. Only returns visits that are not QAQC visits}
\item{TRUE}{Returns all visits, including QAQC visits}}}

\item{locType}{Allows you to only include plots that are part of the GRTS sample design or
include all plots, such as deer exclosures.
\describe{
\item{"VS"}{Only include plots that are part of the Vital Signs GRTS sample design}
\item{"all"}{Include all plots, such as plots in deer exclosures or test plots.}}}

\item{eventType}{Allows you to include only complete sampling events or all sampling events
\describe{
\item{"complete"}{Default. Only include sampling events for a plot that are complete.}
\item{"all}{Include all plot events with a record in tblCOMN.Event, including plots missing most of the data
associated with that event (eg COLO-380-2018). This feature is currently hard-coded in the function.}}}

\item{status}{Filter by live, dead, or all. Acceptable options are:
\describe{
\item{"active"}{Includes all trees with an active monitoring status, including "DF".}
\item{"live"}{live trees only}
\item{"dead"}{dead trees only}
\item{"all"}{Includes all trees with any status, including excluded or missing.}
}}

\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
\item{"invasive"}{Returns species on the Indicator Invasive List}
}}

\item{canopyForm}{Allows you to filter on species growth form
\describe{
\item{"all"}{Default. Returns all species, including low canopy species.}
\item{"canopy"}{Returns canopy-forming species only.}
}}

\item{valueType}{Allows you to return cover class midpoints (numeric) or cover class ranges (text)
\describe{
\item{"all"}{Default. Returns columns for midpoint and cover classes for each quad}
\item{"midpoint"}{Default. Returns numeric cover class midpoints, with Pct prefix.}
\item{"classes"}{Returns the text cover class definitions, with Txt prefix.}
\item{"averages"}{Returns only the plot-level average cover and percent frequency.}
}}
}
\value{
Returns a dataframe with a row for each species/visit combination for quadrat data
}
\description{
This function compiles sapling data collected in microplots, with a record for each sapling measured,
its Status, and DBH. If no live saplings were observed, returns "None present" for ScientificName and 0 for Count. If a
record has a blank ScientificName and associated data, it means it's a missing value. These are rare, but mostly occur
in data <2011.
}
\examples{
importData()
# compile sapling data for invasive species in VAFO for all years
VAFO_quads <- joinQuadSaplings(park = 'VAFO', speciesType = 'invasive')

# compile native saplings only for all parks in cycle 3
native_quads <- joinQuadSaplings(speciesType = 'native', from = 2015, to = 2018)

}
