% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinQuadSeedlings.R
\name{joinQuadSeedlings}
\alias{joinQuadSeedlings}
\title{joinQuadSeedlings: compiles seedling data collected in quadrats}
\usage{
joinQuadSeedlings(
  park = "all",
  from = 2007,
  to = 2021,
  QAQC = FALSE,
  panels = 1:4,
  locType = c("VS", "all"),
  eventType = c("complete", "all"),
  speciesType = c("all", "native", "exotic", "invasive"),
  canopyForm = c("all", "canopy"),
  valueType = c("all", "midpoint", "classes"),
  ...
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"APCO"}{Appomattox Court House NHP only}
\item{"ASIS"}{Assateague Island National Seashore}
\item{"BOWA"}{Booker T. Washington NM only}
\item{"COLO"}{Colonial NHP only}
\item{"FRSP"}{Fredericksburg & Spotsylvania NMP only}
\item{"GETT"}{Gettysburg NMP only}
\item{"GEWA"}{George Washington Birthplace NM only}
\item{"HOFU"}{Hopewell Furnace NHS only}
\item{"PETE"}{Petersburg NBP only}
\item{"RICH"}{Richmond NB only}
\item{"SAHI"}{Sagamore Hill NHS only}
\item{"THST"}{Thomas Stone NHS only}
\item{"VAFO"}{Valley Forge NHP only}}}

\item{from}{Year to start analysis, ranging from 2007 to current year}

\item{to}{Year to stop analysis, ranging from 2007 to current year}

\item{QAQC}{Allows you to remove or include QAQC events.
\describe{
\item{FALSE}{Default. Only returns visits that are not QAQC visits}
\item{TRUE}{Returns all visits, including QAQC visits}}}

\item{panels}{Allows you to select individual panels from 1 to 4. Default is all 4 panels (1:4).
If more than one panel is selected, specify by c(1, 3), for example.}

\item{locType}{Allows you to only include plots that are part of the GRTS sample design or
include all plots, such as deer exclosures.
\describe{
\item{"VS"}{Only include plots that are part of the Vital Signs GRTS sample design}
\item{"all"}{Include all plots, such as plots in deer exclosures or test plots.}}}

\item{eventType}{Allows you to include only complete sampling events or all sampling events
\describe{
\item{"complete"}{Default. Only include sampling events for a plot that are complete.}
\item{"all}{Include all plot events with a record in tblCOMN.Event, including plots missing most of the data
associated with that event (eg ACAD-029.2010). This feature is currently hard-coded in the function.}}}

\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
\item{"invasive"}{Returns species on the Indicator Invasive List}
}}

\item{canopyForm}{Allows you to filter on species growth form
\describe{
\item{"all"}{Default. Returns all species, including low canopy species.}
\item{"canopy"}{Returns canopy-forming species only}
}}

\item{valueType}{Allows you to return cover class midpoints (numeric) or cover class ranges (text)
\describe{
\item{"all"}{Default. Returns columns for midpoint and cover classes for each quad}
\item{"midpoint"}{Default. Returns numeric cover class midpoints, with Pct prefix.}
\item{"classes"}{Returns the text cover class definitions, with Txt prefix.}
}}

\item{...}{Other arguments passed to function.}
}
\value{
Returns a dataframe with a row for each species/visit combination for quadrat data
}
\description{
This function combines seedling data collected in quadrats. If no seedlings were observed, returns
"None present" for ScientificName and 0 for seedling densities. If a record has a blank ScientificName and associated
data, it means it's a missing value. These are rare, but mostly occur in data <2011. Must run importData first. For
quadrat-specific notes, use the joinQuadNotes() function. Note that starting in 2019, all woody species were added to the
indicator list and are summarized in the joinQuadSpecies() function. Seedling cover summarized here only includes tree
seedlings that are seedling size (e.g., doesn't include overhanging branches). After cycle 4 is completed, seedling % cover
will be phased out. For more information about protocol changes, refer to the Summary of Major Protocol Changes document
for the MIDN forest protocol located in the Long-Term Forest Monitoring Protocol IRMA Project:
   https://irma.nps.gov/Datastore/Reference/Profile/2189101.
}
\examples{
\dontrun{
importData()
# compile seedling data for invasive species in VAFO for all years
VAFO_quads <- joinQuadSeedlings(park = 'VAFO', speciesType = 'invasive')

# compile native seedlings only for all parks in cycle 3
native_quads <- joinQuadSeedlings(speciesType = 'native', from = 2015, to = 2018)
}

}
