% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinSoilSampleData.R
\name{joinSoilSampleData}
\alias{joinSoilSampleData}
\title{joinSoilSampleData: compile corrected soil sample data.}
\usage{
joinSoilSampleData(
  park = "all",
  from = 2010,
  to = 2013,
  QAQC = FALSE,
  panels = 1:4,
  locType = c("VS", "all"),
  last_lab_year = 2014,
  ...
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"APCO"}{Appomattox Court House NHP only}
\item{"BOWA"}{Booker T. Washington NM only}
\item{"FRSP"}{Fredericksburg & Spotsylvania NMP only}
\item{"GETT"}{Gettysburg NMP only}
\item{"GEWA"}{George Washington Birthplace NM only}
\item{"HOFU"}{Hopewell Furnace NHS only}
\item{"PETE"}{Petersburg NBP only}
\item{"RICH"}{Richmond NB only}
\item{"SAHI"}{Sagamore Hill NHS only}
\item{"THST"}{Thomas Stone NHS only}
\item{"VAFO"}{Valley Forge NHP only}}}

\item{from}{Year to start analysis, ranging from 2010 to 2013.}

\item{to}{Year to stop analysis, ranging from 2010 to 2013.}

\item{QAQC}{Allows you to remove or include QAQC events.
\describe{
\item{FALSE}{Default. Only returns visits that are not QAQC visits}
\item{TRUE}{Returns all visits, including QAQC visits}}}

\item{panels}{Allows you to select individual panels from 1 to 4. Default is all 4 panels (1:4).
If more than one panel is selected, specify by c(1, 3), for example.}

\item{locType}{Allows you to only include plots that are part of the GRTS sample design or
include all plots, such as deer exclosures.
\describe{
\item{"VS"}{Only include plots that are part of the Vital Signs GRTS sample design}
\item{"all"}{Include all plots, such as plots in deer exclosures or test plots.}}}

\item{last_lab_year}{The most recent year lab analyses have been completed for. This will allow non-lab QCed horizon data
to be returned for years following. Otherwise, only QCed data horizon data are returned.}
}
\value{
returns a dataframe containing each plot and visit with soil sample data.Tot_Samp_cm is the total depth of
O and A sampled. Litter is not included in total depth calculation. Note soil chemistry data are typically a year behind
plot data, so corrected soil horizon depths will also be a year behind. Soil horizon data not yet QCed by lab data are
indicated by the Lab_QC (T/F) column. Plots that weren't sampled during a given cycle are not returned. Horizon depths
are averaged across samples of the same horizon type.
}
\description{
This function verifies whether O and A horizons were named corrected based on % Total Carbon (O = TC >= 20%) using
the joinSoilLabData() function, then compiles average sample depth by horizon. Must run importData first. Only works for complete
visits and plots that haven't been abandoned. Note that Earthworms are summarized in joinStandData(). Note that soils have only
been collected once per plot from 2010 to 2014, and have not been sampled in ASIS or COLO.
}
\examples{
importData() #default imports
# join horizon depth data for GETT
soil_GETT_O <- joinSoilSampleData(park = 'GETT', from = 2010, to = 2013)

}
