% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinTreeVineSpecies.R
\name{joinTreeVineSpecies}
\alias{joinTreeVineSpecies}
\title{joinTreeVineSpecies: compiles trees with vine conditions and the vine species}
\usage{
joinTreeVineSpecies(
  park = "all",
  from = 2007,
  to = 2021,
  QAQC = FALSE,
  locType = c("VS", "all"),
  panels = 1:4,
  speciesType = c("all", "native", "exotic"),
  dist_m = NA,
  ...
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"APCO"}{Appomattox Court House NHP only}
\item{"ASIS"}{Assateague Island National Seashore}
\item{"BOWA"}{Booker T. Washington NM only}
\item{"COLO"}{Colonial NHP only}
\item{"FRSP"}{Fredericksburg & Spotsylvania NMP only}
\item{"GETT"}{Gettysburg NMP only}
\item{"GEWA"}{George Washington Birthplace NM only}
\item{"HOFU"}{Hopewell Furnace NHS only}
\item{"PETE"}{Petersburg NBP only}
\item{"RICH"}{Richmond NB only}
\item{"SAHI"}{Sagamore Hill NHS only}
\item{"THST"}{Thomas Stone NHS only}
\item{"VAFO"}{Valley Forge NHP only}}}

\item{from}{Year to start analysis, ranging from 2007 to current year}

\item{to}{Year to stop analysis, ranging from 2007 to current year}

\item{QAQC}{Allows you to remove or include QAQC events.
\describe{
\item{FALSE}{Default. Only returns visits that are not QAQC visits}
\item{TRUE}{Returns all visits, including QAQC visits}}}

\item{locType}{Allows you to only include plots that are part of the GRTS sample design or include all plots,
such as deer exclosures
\describe{
\item{"VS"}{Only include plots that are part of the Vital Signs GRTS sample design}
\item{"all"}{Include all plots, such as plots in deer exclosures or test plots.}}}

\item{panels}{Allows you to select individual panels from 1 to 4. Default is all 4 panels (1:4).
If more than one panel is selected, specify by c(1, 3), for example.}

\item{speciesType}{Allows you to filter vine species on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
}}

\item{dist_m}{Filter trees by a distance that is less than or equal to the specified distance in meters
of the tree to the center of the plot. If no distance is specified, then all trees will be selected. For
example, to select an area of trees that is 100 square meters in area, use a distance of 5.64m.}

\item{...}{Other arguments passed to function.}
}
\value{
returns a data frame for every tree visit with at least one vine condition recorded. Trees
without a vine condition are not returned.
}
\description{
This function compiles vine species data for live trees only. Must run importData first.
Abandoned plots and incomplete visits are excluded from function.
}
\examples{
\dontrun{
importData()
# compile vine data for all parks in cycle 3, excluding QAQC visits
vines_c3 <- joinTreeVineSpecies(from = 2015, to = 2018, QAQC = FALSE)

# compile exotic vine data for all parks and years
vines_exo <- joinTreeVineSpecies(speciesType = 'exotic')

# compile vine data for GETT in 2019, including QAQC visits
GETT_vines <- joinTreeVineSpecies(park = "GETT", from = 2019, to = 2019,
                                 QAQC = TRUE)
}
}
