% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTreeGrowth.R
\name{plotTreeGrowth}
\alias{plotTreeGrowth}
\title{plotTreeGrowth: plot individual tree growth and mortality}
\usage{
plotTreeGrowth(
  park = "all",
  from = 2007,
  to = as.numeric(format(Sys.Date(), "\%Y")),
  locType = c("VS", "all"),
  panels = 1:4,
  subunit = NA,
  plotName = NA_character_,
  species = NA_character_,
  elev_mort = FALSE,
  exc_wind = FALSE,
  title = TRUE
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"APCO"}{Appomattox Court House NHP only}
\item{"ASIS"}{Assateague Island National Seashore}
\item{"BOWA"}{Booker T. Washington NM only}
\item{"COLO"}{Colonial NHP only}
\item{"FRSP"}{Fredericksburg & Spotsylvania NMP only}
\item{"GETT"}{Gettysburg NMP only}
\item{"GEWA"}{George Washington Birthplace NM only}
\item{"HOFU"}{Hopewell Furnace NHS only}
\item{"PETE"}{Petersburg NBP only}
\item{"RICH"}{Richmond NB only}
\item{"SAHI"}{Sagamore Hill NHS only}
\item{"THST"}{Thomas Stone NHS only}
\item{"VAFO"}{Valley Forge NHP only}}}

\item{from}{Year to start analysis, ranging from 2006 to current year}

\item{to}{Year to stop analysis, ranging from 2006 to current year}

\item{locType}{Allows you to only include plots that are part of the GRTS sample design or include all plots,
such as deer exclosures
\describe{
\item{"VS"}{Only include plots that are part of the Vital Signs GRTS sample design}
\item{"all"}{Include all plots, such as plots in deer exclosures or test plots.}}}

\item{panels}{Allows you to filter individual panels from 1 to 4. Default is all 4 panels (1:4).
If more than one panel is selected, specify by c(1, 3), for example.}

\item{subunit}{Filter park data by subunit. Must match units within a park. Valid inputs:
COLO: "JAMES", "YORK"
FRSP: "CHWILD", "FRED", "SPOT"
PETE: "EAST", "FIVE"}

\item{plotName}{Allows you to select a specific plot to run function for. Value inputs are "PARK-###", like "FRSP-001".
If no plot name is specified, or multiple plots are specified, the function will facet on plotName.}

\item{species}{Allows you to filter on species}

\item{elev_mort}{Logical. If TRUE, only includes plots that have experienced at least 5 percent mortality between a cycle and over
the time interval specified. If FALSE (default), includes all plots with at least one live tree of the species
specified.}

\item{exc_wind}{Logical. If TRUE, excludes trees where windthrow was the likely cause of mortality, due to
a dead broken or dead fallen status the first time a tree was observed to be dead. If FALSE (default), includes
trees that died from windthrow or other causes.}

\item{title}{Logical. If TRUE (default) will include a title for the graph, which either shows park, species, or subunit,
depending on the arguments specified.}
}
\value{
Returns a plot of tree growth rates over time, color coded by rate and whether a tree died.
}
\description{
This function plots individual tree growth over time for all trees that were alive on their
first record in the plot. Trees that started out as dead are not plotted. Each line on a graph represents
the DBH of a live tree up until the first time it is classified as dead. Lines are color coded based
on growth between visits to indicate shrinkage, slowing of growth, and/or fast growth. Each tick or X on the
figure is a sample event for a given tree.
}
\examples{
\dontrun{
importData()

# Make plot of loblolly pine in PETE subunit Five Forks.
plotTreeGrowth(park = "PETE", subunit = "FIVE", species = "Pinus taeda")

# Make plot of loblolly pine in PETE subunit Eastern and Western Front.
plotTreeGrowth(park = "PETE", subunit = "EAST", species = "Pinus taeda")

# Plot ash in GETT
plotTreeGrowth(park = "GETT", species = c("Fraxinus americana", "Fraxinus pennsylvanica", "Fraxinus nigra", "Fraxinus"))

# Specific plots
plotTreeGrowth(plotName = "PETE-269")

# Plot beech for GETT with vertical line where EAB was detected by crew
frax <- c("Fraxinus americana", "Fraxinus pennsylvanica", "Fraxinus nigra", "Fraxinus")
plotTreeGrowth(park = "GETT", species = frax) +
  geom_vline(xintercept = 2017, lty = 'dashed')

# Plot all species of oaks in RICH
oaks <- joinTreeData(park = "RICH") |> filter(grepl("Quercus", ScientificName)) |> 
  select(ScientificName) |> distinct() 
plotTreeGrowth(park = "RICH", species = oaks$ScientificName)

# Plot COLO plots in Jamestown unit
plotTreeGrowth(park = "COLO", subunit = "JAMES")

# View RICH plots with elevated mortality
plotTreeGrowth(park = "RICH", elev_mort = T)

# View RICH plots with elevated mortality not from windthrow
plotTreeGrowth(park = "RICH", elev_mort = T, exc_wind = T)

}

}
