% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_FVM.R
\name{theme_FVM}
\alias{theme_FVM}
\title{theme_FVM: custom ggplot2 theme for forestMIDN}
\usage{
theme_FVM()
}
\value{
This function must be used in conjunction with a ggplot object
}
\description{
This function customizes a theme for plotting MIDN forest data, including removing the
default panel grids from ggplot2 figures.
}
\examples{
\dontrun{
library(forestMIDN)
importData()

library(ggplot2)

 dbh_dist <- sumTreeDBHDist(park = "GETT", status = "live", from = 2022, to = 2025) |>
   tidyr::pivot_longer(cols = starts_with("dens"), names_to = "size_class", values_to = "density") |>
   arrange(Plot_Name, SampleYear, size_class)

 ggplot(dbh_dist, aes(x = size_class, y = density)) +
        geom_bar(stat = 'identity', fill = 'dimgrey', position = position_dodge()) +
        labs(x='Live tree DBH size classes (cm)', y='stems/ha') +
        scale_x_discrete(labels=c('10 - 19.9', '20 - 29.9', '30 - 39.9', '40 - 49.9','50 - 59.9',
                                  '60 - 69.9', '70 - 79.9', '80 - 89.9', '90 - 99.9', '100+')) +
        theme(axis.text.x = element_text(angle = 90, vjust = 0.5)) +
        theme_FHM()

}
}
